requirements.txt
python-dateutil==2.5.3
requests==2.10.0
schedule==0.3.2
six==1.10.0
-e git+https://github.com/nikitaborisov/slackbot.git@spraibot#egg=slackbot
slacker==0.9.18
SQLAlchemy==1.0.13
websocket-client==0.37.0
dateparser>=0.5.0
